Cache pondering results
